Finished Method used to calculate difference between two genomes, eliminated some implicit conversions, code now uses size_t (in this case unsigned long long) a lot more often then the previous unsigned int

+++ JNF_NEAT/gene.cpp
@@ -1,6 +1,6 @@
 #include "gene.h"
 
-unsigned int Gene::numberOfExistingGenes = 0U;
+size_t Gene::numberOfExistingGenes = 0U;
 
 Gene::Gene()
 {
