+++ JNF_NEAT/gene_examinator.cpp
@@ -1,7 +1,7 @@
 #include "gene_examinator.h"
 #include <algorithm>
 
-unsigned int GeneExaminator::GetNumberOfNeuronsInGenes(const std::vector<Gene>& genes)
+size_t GeneExaminator::GetNumberOfNeuronsInGenes(const std::vector<Gene>& genes)
 {
 	auto CompareToNeuron = [](const Gene& lhs, const Gene& rhs) {
 		return lhs.to < rhs.to;
